Hack to improve the repl workflow, performance optimizations needed
By default, none of the JsiiObjects that get created are garbage
collected. As a result, the performance degrades over the lifecycle of
a repl session.

To counteract this, a new runtime is created for every evaluation of
defapp. This only takes ~1 second but prevents the degredation of
performance over time.