Add browse function to core namespace
The first version of the API allowed you to navigate to documentation
by calling a class or instance with `:cdk/browse` as the only
argument. However, this approach was not flexible as it required a
class or object instance to navigate. By introducing a top-level API
function we can make it easy to navigate to the root documentation and
to navigate by string values. This makes it easier to dive into the
documentation without a starting point.