Account for zero arg methods and improve interned fn error messages
This commit accounts for methods which require zero args by
defaulting to an empty vector when no arguments are
provided. Previously a nil would be passed, which would cause some
versions of jackson-databind to throw a class cast exception. Other
version of jackson-databind seem to handle the nil case for
us. Regardless we want this to be as reliable as possible, so we now
check for nil and pass an empty vector in this case. If there is an
exception that occurs when calling these interned functions, it will now
be propagated with all of the information needed to debug the issue.