Allow aliasing with `:as` to rename imports
There are classes with conflicting names in CDK. For example
`@aws-cdk/core.CfnParameter` and `@aws-cdk/aws-ssm.CfnParameter`. This
causes a conflict that is currently unresolvable.

This change adds the ability to change the name of an import with an
`:as` flag.

Example:

``` clojure
(import ["@aws-cdk/aws-ssm"
         CfnParameter :as CfnSsmParameter])
```

This would alias the `@aws-cdk/aws-ssm.CfnParameter` namespace and
refer the `@aws-cdk/aws-ssm.CfnParameter` class to the symbol
`CfnSsmParameter` in the current namespace.