Fix jsii runtime upgrade issue and improve assembly interface
jsii changed the spec of interfaces with the upgrade to 0.20.x. Prior
to the upgrade, jsii returned references to an interface named with
the protocol name in the reference (e.g. `{"$jsii.byref":
"@aws-cdk/aws-ssm.IParameter@10001"}`). After the change, jsii returns
a non-interface class as the object reference and a list of interfaces
the instance implements.

This broke our specs since the we weren't looking for the interfaces
key and the returned object reference no longer implemented the
required interface (since that infromation is now contained in the
interfaces key).

This change fixes this issue while remaining backwards compatibility
with previous jsii runtimes by conditionally looking for the
interfaces key and using the results to check the satisfaction of the
interface when the key is present.

To clean up the implementation, the assembly namespace no longer tries
to expand type information by merging interfaces and base classes in
its object heirarchy. Instead, it presents an interface to retrieve
`methods`, `properties`, `interfaces` and the `class-heirarchy`. These
functions take into account any base classes and interfaces that need
to be accounted for rather than doing the rollup ahead of time.

In working through this issue a few other issues were identified and
fixed:
- static properties are now only rendered on the JsiiClass printer
- instance properties are now only rendered on the JsiiObject printer
- protected properties are no longer printed
- protected methods are no longer interned