Don't partial impl predicates in specs
The predicates in the impl namespace check for instances of the types
definied in that namespace and the spec namespace caches the spec
definitions as they are loaded lazily. When partial was used, this
caused the currently defined type to be used for `instance?` checks
which would be replaced by a new type when `deftype` was re-evaluated
in impl. This would cause all of the cached specs to break since the
type they were checking is no longer the type being created. This
change removes this behavior by removing `partial` and forcing the
predicates to be resolved as they are used rather than being cached
when they are generated.