Remove valid property guards.
Another consequence of the change to the way that JSii implements interfaces,
first referenced in f61eb15897cf025a6d9e9f5880d0ceb7ddd5b987.  These guards to
validate the properties don't correctly handle the case where the property is
defined by the interface that a class implements, not the class itself (as is
the case with [GitHubSourceAction]

cdk-clj incorrectly throws an "Invalid property" errors because its looking for
the property in the wrong place.

The decision for now is to have cdk-clj behave more as a pass-through, and allow
the JSii errors to surface, rather than try to be smart and have to keep up with
these kinds of fundamental architectural changes in JSii.

[GitHubSourceAction]: https://docs.aws.amazon.com/cdk/api/latest/docs/@aws-cdk_aws-codepipeline-actions.GitHubSourceAction.html